Reorganize Repo
Reorganize files and adjust docs to match.
